# backend/src/docs/swagger.yaml
basePath: /
definitions:
  AntifroadKeyGetHandler.ResponseBody:
    properties:
      key:
        example: "12345"
        type: string
    type: object
...
  ResponseError:
    properties:
      error:
        example: something went wrong
        type: string
    type: object
...
info:
  contact: {}
  title: Blindtyping
  version: "0.1"
paths:
  /antifroad/key:
    get:
      consumes:
      - application/json
      description: Получить актуальный антофрод-ключ. Ручка используется только внутри
        контура blindtyping.
      parameters:
      - description: Пароль для авторизации в модуле антифрода
        example: "12345"
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Антифрод-ключ
          schema:
            $ref: '#/definitions/AntifroadKeyGetHandler.ResponseBody'
        "400":
          description: Отсутствуем пароль
          schema:
            $ref: '#/definitions/ResponseError'
        "401":
          description: Неправильный пароль
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: Ключ не найден (вообще такого быть не должно)
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Получить актуальный антофрод-ключ
      tags:
      - Antifroad
  /antifroad/rotate-keys:
    post:
      consumes:
      - application/json
      description: Ротировать ключи. Ручка используется только внутри контура blindtyping.
      parameters:
      - description: Пароль для авторизации в модуле антифрода
        example: "12345"
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Отсутствуем пароль
          schema:
            $ref: '#/definitions/ResponseError'
        "401":
          description: Неправильный пароль
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Ротировать ключи (служебная ручка)
      tags:
      - Antifroad
  /auth/{provider}:
    get:
      consumes:
      - application/json
      description: Initiates the OAuth authentication process by redirecting the user
        to the provider's login page.
      parameters:
      - description: OAuth provider name
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User already authenticated
        "400":
          description: Missing provider or invalid input
          schema:
            $ref: '#/definitions/ResponseError'
      summary: OAuth Login Redirect
      tags:
      - Auth
  /auth/{provider}/callback:
    get:
      consumes:
      - application/json
      description: |-
        Обрабатывает callback-запрос от OAuth провайдера. Завершает процесс аутентификации и выполняет перенаправление на соответствующий фронтенд URL в зависимости от результата.

        **Сценарии перенаправления:**
        - Успешная аутентификация существующего пользователя → `logged_in_url`
        - Первая аутентификация нового пользователя → `registration_url`
        - Ошибка аутентификации → `error_redirection_url?status=STATUS`

        **Коды ошибок:**
        - `1` - Email уже занят другие аккаунтом

        **Особенности:**
        - Устанавливает authentication cookies при успешной аутентификации или registration_token при необходимости регистрации
        - Всегда возвращает HTTP 308 (Permanent Redirect)
        - Параметр `status` передается только в случае ошибки
      parameters:
      - description: Название OAuth провайдера (google или github)
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "308":
          description: Перенаправление на фронтенд
      summary: OAuth Callback Handler
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Delete session (refresh token), add access token to block list
        for some time and delete tokens from cookies
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "403":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: User session not found
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Server side error
          schema:
            $ref: '#/definitions/ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - Auth
...
swagger: "2.0"